KTOR-9703 Optimization for String functions
